requirements.txt
tgtg==0.15.0
python-dotenv==1.0.0